requirements.txt
# Python dependencies
# Install via 
# pip install -r requirements.txt
pyyaml
pyvcd

